OrderStatusInfo.kt pojo fix

+++ ipbandroidapi/src/main/java/com/progressterra/ipbandroidapi/api/ipbDeliveryService/models/delivery/OrderStatusInfo.kt
@@ -4,33 +4,33 @@ import com.google.gson.annotations.SerializedName
 
 data class OrderStatusInfo(
 
-	@SerializedName("rdCreationDate")
-	val rdCreationDate: String? = null,
+    @SerializedName("rdCreationDate")
+    val rdCreationDate: String? = null,
 
-	@SerializedName("rdStatusDate")
-	val rdStatusDate: String? = null,
+    @SerializedName("rdStatusDate")
+    val rdStatusDate: String? = null,
 
-	@SerializedName("rdDeliveryDate")
-	val rdDeliveryDate: String? = null,
+    @SerializedName("rdDeliveryDate")
+    val rdDeliveryDate: String? = null,
 
-	@SerializedName("rfDeliveryService")
-	val rfDeliveryService: String? = null,
+    @SerializedName("rfDeliveryService")
+    val rfDeliveryService: String? = null,
 
-	@SerializedName("rfOrderStatusDescription")
-	val rfOrderStatusDescription: String? = null,
+    @SerializedName("rfOrderStatusDescription")
+    val rfOrderStatusDescription: String? = null,
 
-	@SerializedName("rdNumberPosition")
-	val rdNumberPosition: Int? = null,
+    @SerializedName("rdNumberPosition")
+    val rdNumberPosition: Int? = null,
 
-	@SerializedName("rdAddress")
-	val rdAddress: String? = null,
+    @SerializedName("rdAddress")
+    val rdAddress: String? = null,
 
-	@SerializedName("rdOrderNumber")
-	val rdOrderNumber: String? = null,
+    @SerializedName("rdOrderNumber")
+    val rdOrderNumber: String? = null,
 
-	@SerializedName("rdAmount")
-	val rdAmount: Int? = null,
+    @SerializedName("rdAmount")
+    val rdAmount: Double? = null,
 
-	@SerializedName("rfOrderStatus")
-	val rfOrderStatus: String? = null
+    @SerializedName("rfOrderStatus")
+    val rfOrderStatus: String? = null
 )
